moved make_shared_array to accelerator.hpp
